♻️ merge auth.test.ts into server.test.ts
